Move suppressed exceptions to exceptions
