+++ perun/collect/complexity/symbols.py
@@ -14,7 +14,7 @@ import dataclasses
 
 import perun.utils as utils
 
-from perun.utils import exceptions, helpers
+from perun.utils import exceptions
 
 # Symbol table columns constants
 _SYMTABLE_NAME_COLUMN = 8
@@ -210,7 +210,7 @@ def _dismantle_symbols(symbol_map: dict[str, str]) -> dict[str, PrototypeParts]:
     specification_map = dict()
     for key in symbol_map.keys():
         # Process each symbol from the map
-        with helpers.SuppressedExceptions(exceptions.UnexpectedPrototypeSyntaxError):
+        with exceptions.SuppressedExceptions(exceptions.UnexpectedPrototypeSyntaxError):
             specification_map[key] = _process_symbol(symbol_map[key])
     return specification_map
