+++ perun/collect/trace/optimizations/resources/perun_dynamic_stats.py
@@ -5,8 +5,7 @@
 import perun.logic.stats as stats
 import perun.logic.temp as temp
 import perun.collect.trace.optimizations.dynamic_stats as dyn_stats
-from perun.utils.helpers import SuppressedExceptions
-from perun.utils.exceptions import StatsFileNotFoundException
+from perun.utils.exceptions import StatsFileNotFoundException, SuppressedExceptions
 
 
 def extract(stats_name, reset_cache, **_):
