+++ perun/utils/helpers.py
@@ -10,7 +10,11 @@ from typing import Optional, Any, Iterable, Callable, Literal, TYPE_CHECKING
 
 import perun.postprocess.regression_analysis.tools as tools
 
-from perun.utils.exceptions import SignalReceivedException, NotPerunRepositoryException
+from perun.utils.exceptions import (
+    SignalReceivedException,
+    NotPerunRepositoryException,
+    SuppressedExceptions,
+)
 
 if TYPE_CHECKING:
     import traceback
@@ -151,41 +155,6 @@ def uid_getter(uid: tuple[str, Any]) -> int:
     return uid_priority.get(
         uid[0], int("".join(map(str, map(lambda x: x + max_value, map(ord, uid[0])))))
     )
-
-
-class SuppressedExceptions:
-    """Context manager class for code blocks that need to suppress / ignore some exceptions
-    and simply continue in the execution if those exceptions are encountered.
-
-    :ivar list exc: the list of exception classes that should be ignored
-    """
-
-    __slots__ = ["exc"]
-
-    def __init__(self, *exception_list: type[Exception]) -> None:
-        """
-        :param exception_list: the exception classes to ignore
-        """
-        self.exc = exception_list
-
-    def __enter__(self) -> "SuppressedExceptions":
-        """Context manager entry sentinel, no set up needed
-
-        :return object: the context manager class instance, shouldn't be needed
-        """
-        return self
-
-    def __exit__(self, exc_type: str, exc_val: Exception, exc_tb: traceback.StackSummary) -> bool:
-        """Context manager exit sentinel, check if the code raised an exception and if the
-        exception belongs to the list of suppressed exceptions.
-
-        :param type exc_type: the type of the exception
-        :param exception exc_val: the value of the exception
-        :param traceback exc_tb: the traceback of the exception
-        :return bool: True if the encountered exception should be ignored, False otherwise or if
-                      no exception was raised
-        """
-        return isinstance(exc_val, tuple(self.exc))
 
 
 def str_to_plural(count: int, verb: str) -> str:
